added cutom checkbox + moved todo item to own component
