+++ code/package.json
@@ -15,6 +15,7 @@
     "react": "^16.12.0",
     "react-custom-checkbox": "^1.0.7",
     "react-dom": "^16.12.0",
+    "react-icons": "^4.2.0",
     "react-redux": "^7.2.3",
     "react-scripts": "3.2.0",
     "uuid": "^8.3.2"
